+++ indexer.go
@@ -6,8 +6,10 @@ package main
 
 import (
 	"context"
+	"sort"
 	"time"
 
+	"github.com/Masterminds/semver/v3"
 	"github.com/elastic/package-registry/metrics"
 	"github.com/elastic/package-registry/packages"
 )
@@ -44,5 +46,44 @@ func (c CombinedIndexer) Get(ctx context.Context, opts *packages.GetOptions) (pa
 		}
 		packages = packages.Join(p)
 	}
+
+	if !opts.Filter.AllVersions {
+		return latestPackagesVersion(packages), nil
+	}
+
 	return packages, nil
+}
+
+func latestPackagesVersion(source packages.Packages) (result packages.Packages) {
+	sort.Sort(byNameVersion(source))
+
+	current := ""
+	for _, p := range source {
+		if p.Name == current {
+			continue
+		}
+
+		current = p.Name
+		result = append(result, p)
+	}
+
+	return result
+}
+
+type byNameVersion packages.Packages
+
+func (p byNameVersion) Len() int      { return len(p) }
+func (p byNameVersion) Swap(i, j int) { p[i], p[j] = p[j], p[i] }
+func (p byNameVersion) Less(i, j int) bool {
+	if p[i].Name != p[j].Name {
+		return p[i].Name < p[j].Name
+	}
+
+	// Newer versions first.
+	iSemVer, _ := semver.NewVersion(p[i].Version)
+	jSemVer, _ := semver.NewVersion(p[j].Version)
+	if iSemVer != nil && jSemVer != nil {
+		return jSemVer.LessThan(iSemVer)
+	}
+	return p[j].Version < p[i].Version
 }